fix missing workspace_id exception
If the windows is not attached to a workspace, the call to
`get_number()` would return `None`. Once this happen, future accesses to
`window.workspace_id` would trigger an exception.

The proposed fix catch the exception and remove the offending window, as
we can expect that a window not attached to a Workspace cannot be
switched to.

In elementaryOS this behaviour is exhibited with the Plank, the dock
application.

Fixes #1